use single precision

+++ include/NoAdaptiveUSM.hpp
@@ -11,8 +11,8 @@ using namespace std;
 class NoAdaptiveUSM{
 
 public:
-	int generateLoGKernel(int size, double sigma, Ipp32f* pKernel );
-	int noAdaptiveUSM(const Ipp32f* pSrc, Ipp32f* pDst, IppiSize roiSize, double sigma, float lambda, int kernelLen);
+	int generateLoGKernel(int size, float sigma, Ipp32f* pKernel );
+	int noAdaptiveUSM(const Ipp32f* pSrc, Ipp32f* pDst, IppiSize roiSize, float sigma, float lambda, int kernelLen);
 private:
 
 };
